버튼 통합 성공
애니메이션 추가 필요
메일 저장 방법 고려 필요
