+++ mailmoajo2/app/src/main/res/layout/activity_gmail_load.xml
@@ -17,8 +17,8 @@
 <!--</androidx.constraintlayout.widget.ConstraintLayout>-->
 
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="400dp"
-    android:layout_height="500dp"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
     android:orientation="vertical">
 
     <WebView
