+++ mailmoajo2/app/src/main/res/layout/fragment_settings.xml
@@ -19,20 +19,20 @@ android:orientation="vertical">
         android:id="@+id/gmailLoadBtn">
 
     </Button>
-    <Button
-        android:layout_width="100dp"
-        android:layout_height="70dp"
-        android:text="get accesstoken"
-        android:id="@+id/retroBtn">
+<!--    <Button-->
+<!--        android:layout_width="100dp"-->
+<!--        android:layout_height="70dp"-->
+<!--        android:text="get accesstoken"-->
+<!--        android:id="@+id/retroBtn">-->
 
-    </Button>
-    <Button
-        android:layout_width="100dp"
-        android:layout_height="70dp"
-        android:text="get mail list"
-        android:id="@+id/retro2Btn">
+<!--    </Button>-->
+<!--    <Button-->
+<!--        android:layout_width="100dp"-->
+<!--        android:layout_height="70dp"-->
+<!--        android:text="get mail list"-->
+<!--        android:id="@+id/retro2Btn">-->
 
-    </Button>
+<!--    </Button>-->
 
     <TextView
     android:layout_width="wrap_content"
